+++ package.json
@@ -25,13 +25,13 @@
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.25.4",
-    "hint": "^6.1.10",
+    "hint": "^6.1.11",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
-    "stylelint": "13.3.x",
-    "stylelint-config-standard": "20.0.x",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "3.17.x",
+    "stylelint-scss": "^3.21.0",
     "webpack": "^5.70.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4"
